fitjacket/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.0
filelock==3.17.0
idna==3.10
packaging==24.2
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
urllib3==2.4.0
virtualenv==20.29.2
google-generativeai==0.3.2
Django>=5.0
python-dotenv

